doc: client js README grammatical suggestions for clarity
Wording/content questions:
line 61: let's not use a slash (/) to mean and/or, is it `signer`, `wallet`, or both? 
line 84: do we always call the API the `Cosmos API`? or is the correct name the `Cosmos SDK API`? 
all lines: @clockworkgr teach me! I am learning and want to hear if this PR was helpful, or how I can do better.